api/requirements.txt
Flask
Flask-CORS
requests
python-dotenv
openai
